Update request
处理data是stream的情况
